Implement MessengerPlugin as a client/server.
The client and server communicate in two ways. The client communicates with the server using regular requests, and the server communicates with the client using a websocket. If it seems more efficient, the websocket could be used instead of the regular requests, eliminating the second means of communication and putting all communication through the one websocket.